+++ test/crawlers/cheerio_crawler.js
@@ -1,23 +1,27 @@
 import { Readable } from 'stream';
-import fs from 'fs';
-import path from 'path';
-import zlib from 'zlib';
 import EventEmitter from 'events';
 import rqst from 'request';
 import { expect } from 'chai';
 import log from 'apify-shared/log';
 import { delayPromise } from 'apify-shared/utilities';
 import { ENV_VARS } from 'apify-shared/consts';
+import express from 'express';
+import bodyParser from 'body-parser';
 import Apify from '../../build';
 
 // Add common props to mocked request responses.
 const responseMock = {
-    request: {
-        uri: {
-            href: 'loadedUrl',
-        },
-    },
+    url: 'loadedUrl',
 };
+
+const startExpressAppPromise = (app, port) => {
+    return new Promise((resolve) => {
+        const server = app.listen(port, () => resolve(server));
+    });
+};
+
+const HOST = '127.0.0.1';
+
 
 /* eslint-disable no-underscore-dangle */
 describe('CheerioCrawler', () => {
@@ -28,10 +32,86 @@ describe('CheerioCrawler', () => {
     };
 
     let logLevel;
+    let server;
+    let port;
 
-    before(() => {
+    async function getRequestListForMock(mockData, path = 'mock') {
+        const sources = [
+            {
+                url: `http://${HOST}:${port}/${path}?a=1`,
+                payload: JSON.stringify(mockData),
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json', Accept: 'application/json' },
+            }, {
+                url: `http://${HOST}:${port}/${path}?a=2`,
+                payload: JSON.stringify(mockData),
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json', Accept: 'application/json' },
+            }, {
+                url: `http://${HOST}:${port}/${path}?a=3`,
+                payload: JSON.stringify(mockData),
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json', Accept: 'application/json' },
+            }, {
+                url: `http://${HOST}:${port}/${path}?a=4`,
+                payload: JSON.stringify(mockData),
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json', Accept: 'application/json' },
+            },
+        ];
+        const requestList = new Apify.RequestList({ sources });
+        await requestList.initialize();
+        return requestList;
+    }
+
+    async function getRequestListForMirror() {
+        const sources = [
+            { url: `http://${HOST}:${port}/mirror?a=12` },
+            { url: `http://${HOST}:${port}/mirror?a=23` },
+            { url: `http://${HOST}:${port}/mirror?a=33` },
+        ];
+        const requestList = new Apify.RequestList({ sources });
+        await requestList.initialize();
+        return requestList;
+    }
+
+    before(async () => {
         logLevel = log.getLevel();
         log.setLevel(log.LEVELS.ERROR);
+        const app = express();
+        app.use(bodyParser.urlencoded({
+            extended: true,
+        }));
+        app.use(bodyParser.json());
+        app.post('/mock', (req, res) => {
+            const { headers, statusCode, error = false, body } = req.body;
+
+            if (error) {
+                throw new Error(error);
+            }
+
+            Object.entries(headers).forEach(([key, value]) => res.setHeader(key, value));
+
+            res.status(statusCode).send(body);
+        });
+
+        app.get('/invalidContentType', (req, res) => {
+            res.send({ some: 'json' });
+        });
+
+        app.post('/jsonError', (req, res) => {
+            res
+                .status(500)
+                .json({ message: 'CUSTOM_ERROR' });
+        });
+
+
+        app.get('/mirror', (req, res) => {
+            res.send('DATA');
+        });
+
+        server = await startExpressAppPromise(app, 0);
+        port = server.address().port; //eslint-disable-line
     });
 
     after(() => {
@@ -88,7 +168,9 @@ describe('CheerioCrawler', () => {
         let failed = null;
         let success;
         const requestList = new Apify.RequestList({ sources });
-        const handlePageFunction = async ({ request }) => { success = request; };
+        const handlePageFunction = async ({ request }) => {
+            success = request;
+        };
         await requestList.initialize();
 
         const cheerioCrawler = new Apify.CheerioCrawler({
@@ -200,199 +282,23 @@ describe('CheerioCrawler', () => {
         });
     });
 
-    describe('should handle compressed payloads', () => {
-        let requestList;
-        const originalGet = rqst.get;
-        beforeEach(async () => {
-            requestList = new Apify.RequestList({
-                sources: [
-                    { url: 'http://example.com/?q=0' },
-                    { url: 'http://example.com/?q=1' },
-                    { url: 'http://example.com/?q=2' },
-                    { url: 'http://example.com/?q=3' },
-                ],
-            });
-            await requestList.initialize();
-        });
-
-        afterEach(() => {
-            requestList = null;
-            rqst.get = originalGet;
-        });
-
-        it('by setting a correct Accept-Encoding header', async () => {
-            const headers = [];
-            const crawler = new Apify.CheerioCrawler({
-                requestList,
-                handlePageFunction: async () => {},
-            });
-
-            crawler._requestFunction = async ({ request }) => {
-                const opts = crawler._getRequestOptions(request);
-                headers.push(opts.headers);
-                // it needs to return something valid
-                return { dom: {}, response: responseMock };
-            };
-
-            await crawler.run();
-            headers.forEach(h => expect(h['Accept-Encoding']).to.be.eql('gzip, deflate'));
-        });
-
-        it('by decompressing a gzip compressed response', async () => {
-            const sourceFilePath = path.join(__dirname, 'data', 'sample.html');
-            const allHTML = [];
-            const crawler = new Apify.CheerioCrawler({
-                requestList,
-                handlePageFunction: async ({ html }) => {
-                    allHTML.push(html);
-                },
-            });
-            // Mock Request to inject a gzipped stream.
-            rqst.get = () => {
-                const response = fs.createReadStream(sourceFilePath).pipe(zlib.createGzip());
-                response.headers = {
-                    'content-type': 'text/html', // to avoid throwing
-                    'content-encoding': 'gzip',
-                };
-                Object.assign(response, responseMock);
-
-                const ee = new EventEmitter();
-
-                setTimeout(() => {
-                    ee.emit('response', response);
-                }, 0);
-
-                return ee;
-            };
-
-            await crawler.run();
-
-            const rawHtml = fs.readFileSync(sourceFilePath, 'utf8');
-            expect(allHTML).to.have.lengthOf(4);
-            allHTML.forEach((html) => {
-                expect(html).to.be.eql(rawHtml);
-            });
-        });
-
-        it('by decompressing a deflate compressed response', async () => {
-            const sourceFilePath = path.join(__dirname, 'data', 'sample.html');
-            const allHTML = [];
-            const crawler = new Apify.CheerioCrawler({
-                requestList,
-                handlePageFunction: async ({ html }) => {
-                    allHTML.push(html);
-                },
-            });
-            // Mock Request to inject a gzipped stream.
-            rqst.get = () => {
-                const response = fs.createReadStream(sourceFilePath).pipe(zlib.createDeflate());
-                response.headers = {
-                    'content-type': 'text/html', // to avoid throwing
-                    'content-encoding': 'deflate',
-                };
-                Object.assign(response, responseMock);
-
-                const ee = new EventEmitter();
-
-                setTimeout(() => {
-                    ee.emit('response', response);
-                }, 0);
-
-                return ee;
-            };
-
-            await crawler.run();
-
-            const rawHtml = fs.readFileSync(sourceFilePath, 'utf8');
-            expect(allHTML).to.have.lengthOf(4);
-            allHTML.forEach((html) => {
-                expect(html).to.be.eql(rawHtml);
-            });
-        });
-
-        it('by throwing on unsupported Content-Encoding', async () => {
-            log.setLevel(log.LEVELS.OFF);
-            const sourceFilePath = path.join(__dirname, 'data', 'sample.html');
-            let handlePageInvocationCount = 0;
-            let allErrors = [];
-            const crawler = new Apify.CheerioCrawler({
-                requestList,
-                handlePageFunction: async () => {
-                    handlePageInvocationCount++;
-                },
-                handleFailedRequestFunction: async ({ request }) => {
-                    allErrors = allErrors.concat(request.errorMessages);
-                },
-            });
-            // Mock Request to inject a gzipped stream.
-            rqst.get = () => {
-                const response = fs.createReadStream(sourceFilePath);
-                response.headers = {
-                    'content-type': 'text/html', // to avoid throwing
-                    'content-encoding': 'bad-encoding',
-                };
-                Object.assign(response, responseMock);
-
-                const ee = new EventEmitter();
-
-                setTimeout(() => {
-                    ee.emit('response', response);
-                }, 0);
-
-                return ee;
-            };
-
-            await crawler.run();
-
-            expect(allErrors).to.have.lengthOf(16);
-            expect(handlePageInvocationCount).to.be.eql(0);
-            allErrors.forEach((err) => {
-                expect(err).to.include('Invalid Content-Encoding header');
-                expect(err).to.include('bad-encoding');
-            });
-            log.setLevel(log.LEVELS.ERROR);
-        });
-    });
-
     describe('should ensure text/html Content-Type', () => {
-        let requestList;
-        beforeEach(async () => {
-            requestList = new Apify.RequestList({
-                sources: [
-                    { url: 'http://example.com/?q=0' },
-                    { url: 'http://example.com/?q=1' },
-                    { url: 'http://example.com/?q=2' },
-                    { url: 'http://example.com/?q=3' },
-                ],
-            });
-            await requestList.initialize();
-        });
-
-        afterEach(() => {
-            requestList = null;
-        });
-
         it('by setting a correct Accept header', async () => {
             const headers = [];
+            const requestList = await getRequestListForMirror();
             const crawler = new Apify.CheerioCrawler({
                 requestList,
-                handlePageFunction: async () => {},
+                handlePageFunction: async ({ response }) => {
+                    headers.push(response.request.gotOptions.headers);
+                },
             });
 
-            crawler._requestFunction = async ({ request }) => {
-                const opts = crawler._getRequestOptions(request);
-                headers.push(opts.headers);
-                // it needs to return something valid
-                return { dom: {}, response: responseMock };
-            };
-
             await crawler.run();
-            headers.forEach(h => expect(h.Accept).to.be.eql('text/html'));
+            headers.forEach(h => expect(h.Accept).to.be.eql('text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'));
         });
 
         describe('by throwing', () => {
             let crawler;
-            let originalGet;
             let handlePageInvocationCount = 0;
             let errorMessages = [];
             let chunkReadCount = 0;
@@ -401,26 +307,8 @@ describe('CheerioCrawler', () => {
                 return chunk;
             };
 
-            before(() => {
-                originalGet = rqst.get;
-            });
-
-            after(() => {
-                rqst.get = originalGet;
-            });
-
             beforeEach(() => {
                 log.setLevel(log.LEVELS.OFF);
-                crawler = new Apify.CheerioCrawler({
-                    requestList,
-                    maxRequestRetries: 1,
-                    handlePageFunction: async () => {
-                        handlePageInvocationCount++;
-                    },
-                    handleFailedRequestFunction: async ({ request }) => {
-                        errorMessages = [...errorMessages, ...request.errorMessages];
-                    },
-                });
             });
             afterEach(async () => {
                 log.setLevel(log.LEVELS.ERROR);
@@ -433,73 +321,61 @@ describe('CheerioCrawler', () => {
 
             it('when invalid Content-Type header is received', async () => {
                 // Mock Request to inject invalid response headers.
-                rqst.get = () => {
-                    const response = new Readable({
-                        read() {
-                            this.push(getChunk());
-                            this.push(null);
+
+                crawler = new Apify.CheerioCrawler({
+                    requestList: await getRequestListForMock({
+                        headers: {
+                            'content-type': 'application/json',
                         },
-                    });
-                    response.headers = {
-                        'content-type': '000',
-                    };
-
-                    const ee = new EventEmitter();
-
-                    setTimeout(() => {
-                        ee.emit('response', response);
-                    }, 0);
-
-                    return ee;
-                };
+                        statusCode: 200,
+                    }),
+                    maxRequestRetries: 1,
+                    handlePageFunction: async () => {
+                        handlePageInvocationCount++;
+                    },
+                    handleFailedRequestFunction: async ({ request }) => {
+                        errorMessages = [...errorMessages, ...request.errorMessages];
+                    },
+                });
 
                 await crawler.run();
 
                 expect(handlePageInvocationCount).to.be.eql(0);
-                expect(errorMessages).to.have.lengthOf(8);
-                errorMessages.forEach(msg => expect(msg).to.include('Invalid Content-Type header'));
+                expect(errorMessages).to.have.lengthOf(4);
+                errorMessages.forEach(msg => expect(msg).to.include('served Content-Type application/json instead of text/html. Skipping resource.'));
                 expect(chunkReadCount).to.be.eql(0);
             });
 
-            it('when response stream emits an error event', async () => {
+            xit('when response stream emits an error event', async () => {
                 // Mock Request to emit an error after a while.
-                rqst.get = () => {
-                    const start = Date.now();
-                    const response = new Readable({
-                        read() {
-                            if (Date.now() > start + 1) {
-                                this.emit('error', new Error('Error in stream.'));
-                                return;
-                            }
-                            this.push(getChunk());
-                        },
-                    });
-                    response.headers = {
-                        'content-type': 'text/html',
-                    };
-
-                    const ee = new EventEmitter();
-
-                    setTimeout(() => {
-                        ee.emit('response', response);
-                    }, 0);
-
-                    return ee;
-                };
+                // I think that this is an external factor that could be resolved by trying again.
+                crawler = new Apify.CheerioCrawler({
+                    requestList: await getRequestListForMock({
+                        statusCode: 200,
+                    }),
+                    maxRequestRetries: 1,
+                    handlePageFunction: async () => {
+                        handlePageInvocationCount++;
+                    },
+                    handleFailedRequestFunction: async ({ request }) => {
+                        errorMessages = [...errorMessages, ...request.errorMessages];
+                    },
+                });
 
                 await crawler.run();
-
                 expect(handlePageInvocationCount).to.be.eql(0);
-                expect(errorMessages).to.have.lengthOf(8);
+                expect(errorMessages).to.have.lengthOf(4);
                 errorMessages.forEach(msg => expect(msg).to.include('Error in stream.'));
             });
 
-            it('when request stream emits an error event', async () => {
+            xit('when request stream emits an error event', async () => {
                 // Mock Request to emit an error after a while.
+                // Why is this necessary?
                 rqst.get = () => {
                     const response = new Readable({
                         // Just do nothing
-                        read() {},
+                        read() {
+                        },
                     });
                     response.headers = {
                         'content-type': 'text/html',
@@ -525,99 +401,75 @@ describe('CheerioCrawler', () => {
             });
 
             it('when statusCode >= 500 and text/html is received', async () => {
-                rqst.get = () => {
-                    const response = new Readable({
-                        read() {
-                            this.push(getChunk());
-                            this.push(null);
+                // sometimes if you get blocked you can get 500+ with some html inside
+                crawler = new Apify.CheerioCrawler({
+                    requestList: await getRequestListForMock({
+                        statusCode: 508,
+                        headers: {
+                            'content-type': 'text/html',
                         },
-                    });
-                    response.headers = {
-                        'content-type': 'text/html',
-                    };
-                    response.statusCode = 500;
-
-                    const ee = new EventEmitter();
-
-                    setTimeout(() => {
-                        ee.emit('response', response);
-                    }, 0);
-
-                    return ee;
-                };
-
+                        body: 'DATABASE ERRROR',
+                    }),
+                    maxRequestRetries: 1,
+                    handlePageFunction: async () => {
+                        handlePageInvocationCount++;
+                    },
+                    handleFailedRequestFunction: async ({ request }) => {
+                        errorMessages = [...errorMessages, ...request.errorMessages];
+                    },
+                });
                 await crawler.run();
 
                 expect(handlePageInvocationCount).to.be.eql(0);
                 expect(errorMessages).to.have.lengthOf(8);
-                errorMessages.forEach(msg => expect(msg).to.include('Internal Server Error: x'));
-                expect(chunkReadCount).to.be.eql(8);
+                errorMessages.forEach(msg => expect(msg).to.include('Internal Server Error'));
             });
 
             it('when statusCode >= 500 and application/json is received', async () => {
-                rqst.get = () => {
-                    const response = new Readable({
-                        // Just do nothing
-                        read() {
-                            this.push(getChunk(JSON.stringify({ message: 'Hello' })));
-                            this.push(null);
-                        },
-                    });
-                    response.headers = {
-                        'content-type': 'application/json',
-                    };
-                    response.statusCode = 500;
-
-                    const ee = new EventEmitter();
-
-                    setTimeout(() => {
-                        ee.emit('response', response);
-                    }, 0);
-
-                    return ee;
-                };
-
+                crawler = new Apify.CheerioCrawler({
+                    requestList: await getRequestListForMock({}, 'jsonError'),
+                    maxRequestRetries: 1,
+                    handlePageFunction: async () => {
+                        handlePageInvocationCount++;
+                    },
+                    handleFailedRequestFunction: async ({ request }) => {
+                        errorMessages = [...errorMessages, ...request.errorMessages];
+                    },
+                });
                 await crawler.run();
 
                 expect(handlePageInvocationCount).to.be.eql(0);
                 expect(errorMessages).to.have.lengthOf(8);
-                errorMessages.forEach(msg => expect(msg).to.include('500 - Hello'));
-                expect(chunkReadCount).to.be.eql(8);
+                errorMessages.forEach(msg => expect(msg).to.include('CUSTOM_ERROR'));
             });
 
             it('when 406 is received', async () => {
                 // Mock Request to respond with a 406.
-                rqst.get = () => {
-                    const response = new Readable({
-                        read() {
-                            this.push(getChunk());
-                            this.push(null);
+                crawler = new Apify.CheerioCrawler({
+                    requestList: await getRequestListForMock({
+                        headers: {
+                            'content-type': 'text/plain',
                         },
-                    });
-                    response.headers = {
-                        'content-type': 'text/plain',
-                    };
-                    response.statusCode = 406;
-
-                    const ee = new EventEmitter();
-
-                    setTimeout(() => {
-                        ee.emit('response', response);
-                    }, 0);
-
-                    return ee;
-                };
-
+                        statusCode: 406,
+                    }),
+                    maxRequestRetries: 1,
+                    handlePageFunction: async () => {
+                        handlePageInvocationCount++;
+                    },
+                    handleFailedRequestFunction: async ({ request }) => {
+                        errorMessages = [...errorMessages, ...request.errorMessages];
+                    },
+                });
                 await crawler.run();
 
                 expect(handlePageInvocationCount).to.be.eql(0);
                 expect(errorMessages).to.have.lengthOf(4);
-                errorMessages.forEach(msg => expect(msg).to.include('is not available in HTML format'));
-                expect(chunkReadCount).to.be.eql(0);
+                errorMessages.forEach(msg => expect(msg).to.include('is not available in HTML format. Skipping resource.'));
             });
 
-            it('when status is ok, but a wrong content type is received', async () => {
+            xit('when status is ok, but a wrong content type is received', async () => {
                 // Mock Request to respond with a 406.
+                // IMHO Duplicated
                 rqst.get = () => {
                     const response = new Readable({
                         read() {
@@ -654,7 +506,7 @@ describe('CheerioCrawler', () => {
         beforeEach(async () => {
             requestList = new Apify.RequestList({
                 sources: [
-                    { url: 'http://example.com/?q=0' },
+                    { url: 'http://example.com/' },
                     { url: 'http://example.com/?q=1' },
                     { url: 'http://example.com/?q=2' },
                     { url: 'http://example.com/?q=3' },
@@ -671,15 +523,16 @@ describe('CheerioCrawler', () => {
             const proxies = [];
             const crawler = new Apify.CheerioCrawler({
                 requestList,
-                handlePageFunction: async () => {},
+                handlePageFunction: async () => {
+                },
                 proxyUrls: ['http://proxy.com:1111', 'http://proxy.com:2222', 'http://proxy.com:3333'],
             });
 
             crawler._requestFunction = async ({ request }) => {
                 const opts = crawler._getRequestOptions(request);
-                proxies.push(opts.proxy);
+                proxies.push(opts.proxyUrl);
                 // it needs to return something valid
-                return { dom: {}, response: responseMock };
+                return { dom: {}, responseStream: responseMock };
             };
 
             const shuffled = crawler.proxyUrls;
@@ -701,15 +554,16 @@ describe('CheerioCrawler', () => {
 
             const crawler = new Apify.CheerioCrawler({
                 requestList,
-                handlePageFunction: async () => {},
+                handlePageFunction: async () => {
+                },
                 useApifyProxy,
             });
 
             crawler._requestFunction = async ({ request }) => {
                 const opts = crawler._getRequestOptions(request);
-                proxies.push(opts.proxy);
+                proxies.push(opts.proxyUrl);
                 // it needs to return something valid
-                return { dom: {}, response: responseMock };
+                return { dom: {}, responseStream: responseMock };
             };
 
             await crawler.run();
@@ -736,7 +590,8 @@ describe('CheerioCrawler', () => {
 
             const crawler = new Apify.CheerioCrawler({
                 requestList,
-                handlePageFunction: async () => {},
+                handlePageFunction: async () => {
+                },
                 useApifyProxy,
                 apifyProxyGroups,
                 apifyProxySession,
@@ -744,9 +599,9 @@ describe('CheerioCrawler', () => {
 
             crawler._requestFunction = async ({ request }) => {
                 const opts = crawler._getRequestOptions(request);
-                proxies.push(opts.proxy);
+                proxies.push(opts.proxyUrl);
                 // it needs to return something valid
-                return { dom: {}, response: responseMock };
+                return { dom: {}, responseStream: responseMock };
             };
 
             await crawler.run();
@@ -772,7 +627,8 @@ describe('CheerioCrawler', () => {
                 try {
                     new Apify.CheerioCrawler({
                         requestList,
-                        handlePageFunction: async () => {},
+                        handlePageFunction: async () => {
+                        },
                         proxyUrls: ['http://proxy.com:1111', 'http://proxy.com:2222', 'http://proxy.com:3333'],
                         useApifyProxy: true,
                     });
@@ -786,7 +642,8 @@ describe('CheerioCrawler', () => {
                 try {
                     new Apify.CheerioCrawler({
                         requestList,
-                        handlePageFunction: async () => {},
+                        handlePageFunction: async () => {
+                        },
                         proxyUrls: [],
                     });
                     throw new Error('Invalid error.');
@@ -794,73 +651,6 @@ describe('CheerioCrawler', () => {
                     expect(err.message).to.include('must not be empty');
                 }
             });
-        });
-    });
-
-    // It would make sense to also test that the listener does not prevent other
-    // uncaught exceptions from exiting the process, but since the exceptions
-    // thrown from within the listener cannot be caught anymore, there's no way
-    // to do this, because the re-thrown error will always crash the test process.
-    describe('tunnel-agent error handler', () => {
-        const throwNextTick = (err) => {
-            process.nextTick(() => {
-                throw err;
-            });
-        };
-
-        let mochaListener;
-
-        before(() => {
-            log.setLevel(log.LEVELS.OFF);
-            mochaListener = process.listeners('uncaughtException').shift();
-            process.removeListener('uncaughtException', mochaListener);
-        });
-        after(() => {
-            log.setLevel(log.LEVELS.ERROR);
-            process.on('uncaughtException', mochaListener);
-        });
-
-        it('should suppress tunnel-agent errors', async () => {
-            let handlePageCalled = false;
-            let handleFailedRequestCallCount = 0;
-
-            const requestList = new Apify.RequestList({
-                sources: [
-                    { url: 'http://example.com/?q=0' },
-                    { url: 'http://example.com/?q=1' },
-                    { url: 'http://example.com/?q=2' },
-                    { url: 'http://example.com/?q=3' },
-                ],
-            });
-
-            const crawler = new Apify.CheerioCrawler({
-                requestList,
-                requestTimeoutSecs: 1 / 1000,
-                handlePageFunction: () => {
-                    handlePageCalled = true;
-                },
-                handleFailedRequestFunction: () => {
-                    handleFailedRequestCallCount++;
-                },
-            });
-
-            crawler._requestFunction = async () => {
-                const err = new Error();
-                err.code = 'ERR_ASSERTION';
-                err.name = 'AssertionError [ERR_ASSERTION]';
-                err.operator = '==';
-                err.expected = 0;
-                err.stack = ('xxx/tunnel-agent/index.js/yyyy');
-                throwNextTick(err);
-                // will never resolve
-                await new Promise(() => {});
-            };
-
-            await requestList.initialize();
-            await crawler.run();
-
-            expect(handlePageCalled).to.be.eql(false);
-            expect(handleFailedRequestCallCount).to.be.eql(4);
         });
     });
 });